Authentification from the homepage to the admin dashboard complete. UI of the login page needs to be updated to match the rest of the website.
