+++ package.json
@@ -11,7 +11,9 @@
   },
   "dependencies": {
     "@prisma/client": "^6.17.1",
+    "bcrypt": "^6.0.0",
     "framer-motion": "^12.23.24",
+    "jsonwebtoken": "^9.0.2",
     "next": "15.5.4",
     "next-themes": "^0.4.6",
     "react": "19.1.0",
@@ -21,6 +23,8 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
+    "@types/bcrypt": "^6.0.0",
+    "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
