+++ src/app/api/auth/me/route.ts
@@ -1,0 +1,39 @@
+import { NextResponse } from "next/server";
+import { PrismaClient } from "@prisma/client";
+import { cookies } from "next/headers";
+import { verifyToken, COOKIE_NAME } from "@/lib/auth";
+
+const prisma = new PrismaClient();
+
+export async function GET() {
+  try {
+    // Get token from cookies
+    const cookieStore = await cookies();
+    const token = cookieStore.get(COOKIE_NAME)?.value;
+
+    if (!token) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    // Verify JWT
+    const decoded = verifyToken(token);
+    if (!decoded) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    // Fetch admin details
+    const admin = await prisma.admin.findUnique({
+      where: { id: decoded.adminId },
+      select: { id: true, email: true, name: true },
+    });
+
+    if (!admin) {
+      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
+    return NextResponse.json(admin);
+  } catch (err) {
+    console.error("AUTH /me ERROR:", err);
+    return NextResponse.json({ error: "Internal server error" }, { status: 500 });
+  }
+}
